% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{absolute_path}
\alias{absolute_path}
\title{Return absolute path for a given path}
\usage{
absolute_path(path)
}
\arguments{
\item{path}{Path to be made absolute}
}
\value{
Normalize, absolute \code{path} relative to current working directory
}
\description{
Return absolute path for a given path
}
