% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{add_user_to_role}
\alias{add_user_to_role}
\title{Add a user to a role
Run only with operator or admin privileges.  Admin privileges required to grant operator or admin privileges.}
\usage{
add_user_to_role(pkgEnv, con, user_name, role, add_to_inferior_roles = TRUE)
}
\arguments{
\item{user_name}{username of user to add to role}

\item{role}{name of role to add user to.  See \link{show_roles}}

\item{add_to_inferior_roles}{if TRUE (default), also add user to any roles with lesser level than the specified role.  See \link{show_roles}}
}
\description{
Add a user to a role
Run only with operator or admin privileges.  Admin privileges required to grant operator or admin privileges.
}
