% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{add_user_to_role_call}
\alias{add_user_to_role_call}
\title{Wrapper for the SciDB add_user_to_role, with error checking}
\usage{
add_user_to_role_call(con, user_name, role)
}
\description{
Wrapper for the SciDB add_user_to_role, with error checking
}
