% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{check_user_operator_status}
\alias{check_user_operator_status}
\title{Function to check if current user has operator permissions}
\usage{
check_user_operator_status(con, user_name = NULL)
}
\arguments{
\item{con}{a connection object}
}
\description{
Function to check if current user has operator permissions
}
