% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to scidb and return a connection object.}
\usage{
connect(
  pkgEnv,
  username = NULL,
  password = NULL,
  host = NULL,
  port = NULL,
  protocol = "https",
  result_size_limit = 2 * 1048,
  use_test_namespace = FALSE,
  multi_connection_environment = F
)
}
\arguments{
\item{username}{username of user to log in as}

\item{password}{password for the user.  If null, will be requested interactively.}

\item{host}{if NULL, attempt to set automatically from Apache config}

\item{port}{if NULL, attempt to set automatically from Apache config}

\item{protocol}{protocol type}

\item{use_test_namespace}{if TRUE, use a single test namespace defined by the package schema, iff the package has such a namespace}

\item{multi_connection_environment}{if FALSE and the database connection is successful, store the connection object in a global environment variable.  This allows calling package functions without specifying a connection object.  Otherwise, remove connection object from global environment.}

\item{results_size_limit}{Maximum size of any single result from a scidb query over this connection.  Query results larger than the specified size will error.}
}
\value{
a rg_connection object
}
\description{
Connect to scidb and return a connection object.
}
