% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\docType{class}
\name{connection-class}
\alias{connection-class}
\alias{connection}
\title{API Class for Querying and Computing in SciDB}
\description{
API Class for Querying and Computing in SciDB
}

\seealso{
\code{\link{connect}} for connection instructions.
}
