% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\docType{data}
\name{data_type_mapping_r_to_scidb}
\alias{data_type_mapping_r_to_scidb}
\title{Map R to scidb data type}
\format{
An object of class \code{character} of length 4.
}
\usage{
data_type_mapping_r_to_scidb
}
\description{
Map R to scidb data type
}
\keyword{datasets}
