% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\docType{data}
\name{data_type_mapping_scidb_to_r}
\alias{data_type_mapping_scidb_to_r}
\title{Map scidb to R data type}
\format{
An object of class \code{character} of length 5.
}
\usage{
data_type_mapping_scidb_to_r
}
\description{
Map scidb to R data type
}
\keyword{datasets}
