% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{.auto_remove}
\alias{.auto_remove}
\title{Wrap array in \code{.AutoRemove} for cleanup during garbage collection}
\usage{
.auto_remove(array)
}
\arguments{
\item{array}{array to auto remove}
}
\value{
array with .AutoRemove ref added as attribute
}
\description{
This can be used to alleviate gc() issues in arrayop. Generates a new
\code{.AutoRemove} object and attaches it to \code{array} as an attribute.
}
