% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{.AutoRemove}
\alias{.AutoRemove}
\title{Wrapper to auto-remove temporary persistent arrays}
\description{
R6 Class to track arrayop \code{ArrayOpBase} object and auto-remove when
it is garbage collected
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array}}{Reference to \code{ArrayOpBase} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.AutoRemove$new()}}
\item \href{#method-finalize}{\code{.AutoRemove$finalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize new \code{.AutoRemove} with an \code{ArrayOpBase} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.AutoRemove$new(array)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array}}{\code{ArrayOpBase} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{.AutoRemove} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Remove \code{ArrayOpBase} object when garbage collected
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.AutoRemove$finalize()}\if{html}{\out{</div>}}
}

}
}
