% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_file}
\alias{download_file}
\title{Download file using httr}
\usage{
download_file(
  url,
  filename = tempfile(fileext = file_ext(url)),
  ...,
  progress = TRUE,
  ssl_no_verify = FALSE
)
}
\arguments{
\item{url}{Supported schemes are http and https}

\item{filename}{File name to save to (defaults to \code{tempfile()})}

\item{...}{Passed on to \code{httr::config()}}

\item{progress}{If set show download progress bar}

\item{ssl_no_verify}{Turn off SSL certificate and host verification (not recommended)}
}
\value{
File name of saved download
}
\description{
Download file using httr
}
