% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_s3}
\alias{download_s3}
\title{Download S3 object to file}
\usage{
download_s3(object, bucket, opts = NULL, filename = NULL, ..., progress = TRUE)
}
\arguments{
\item{object}{Object name or s3 url}

\item{bucket}{Optional bucket (derived from object if missing)}

\item{opts}{Options passed on to \code{aws.s3::save_object()}}

\item{filename}{File name to save to (defaults to \code{tempfile()})}

\item{...}{Passed as options to \code{aws.s3::save_object()}}

\item{progress}{If set show download progress bar}
}
\value{
File name of saved download
}
\description{
Download S3 object to file
}
