% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{drop_na_columns}
\alias{drop_na_columns}
\title{Drop columns that have NA in all rows}
\usage{
drop_na_columns(df)
}
\description{
Drop columns that have NA in all rows
}
\examples{
{
drop_na_columns(data.frame(a = 1:3, b = NA, c = c('x', 'y', 'z')))
}
}
