% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{drop_user_from_role}
\alias{drop_user_from_role}
\title{Drop a user from a role
Run only with operator or admin privileges.  Admin privileges required to remove operator or admin privileges.}
\usage{
drop_user_from_role(
  pkgEnv,
  con,
  user_name,
  role,
  drop_from_superior_roles = TRUE
)
}
\arguments{
\item{user_name}{username of user to drop from role}

\item{role}{name of role to drop user to.  See \link{show_roles}}

\item{drop_from_superior_roles}{if TRUE (default), also add user to any roles with greater or equal level to the specified role.  See \link{show_roles}}
}
\description{
Drop a user from a role
Run only with operator or admin privileges.  Admin privileges required to remove operator or admin privileges.
}
