% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{drop_user_from_role_call}
\alias{drop_user_from_role_call}
\title{Wrapper for the SciDB drop_user_from_role, with error checking}
\usage{
drop_user_from_role_call(con, user_name, role)
}
\description{
Wrapper for the SciDB drop_user_from_role, with error checking
}
