% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{file_ext}
\alias{file_ext}
\title{Obtain file extention from path}
\usage{
file_ext(url, last_dot_only = FALSE)
}
\arguments{
\item{url}{Path to get extension for}

\item{last_dot_only}{if \code{TRUE} use last \code{.} to determine extension}
}
\value{
File extension for given path
}
\description{
Returns anything after (and including) the first \code{.} found in basename
of path
}
