% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{filter_reducer}
\alias{filter_reducer}
\title{Generate nested filters from a list of expressions and an arrayop object}
\usage{
filter_reducer(array, expressions, case_sensitive = TRUE, ...)
}
\arguments{
\item{array}{arrayop to filter}

\item{expressions}{sequence of expressions (i.e. unevaluated expressions such as
returned by str2lang)}

\item{case_sensitive}{inverted and passed on to .ignore_case of filter()}

\item{...}{extra arguments passed to \emph{all} nested filter calls}
}
\value{
Result of \verb{array$filter(....)$filter(....)....}
}
\description{
Given an object obj and a list of expressions \code{list(a == 1, x > 0, etc.)}, generate
calls \code{obj$filter(a == 1)$filter(x > 0)$filter(....)}.
}
