% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{find_matches_and_return_indices}
\alias{find_matches_and_return_indices}
\title{helper function to report matches between vectors}
\usage{
find_matches_and_return_indices(source, target, match_fcn = base::match)
}
\arguments{
\item{source}{source vector for finding matches from}

\item{target}{target vector in which to find matches}

\item{match_fcn}{a function that does matching between source and target.  Default base::match. Use revealcore::match_df for data frames.}
}
\value{
list(match_res, source_matched_idx, source_unmatched_idx, target_matched_idx)
}
\description{
helper function to report matches between vectors
}
