% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-schema.R
\name{find_namespace}
\alias{find_namespace}
\title{return the namespace of an entity}
\usage{
find_namespace(pkgEnv, entitynm, con = NULL)
}
\arguments{
\item{pkgEnv}{the package environment}

\item{entitynm}{an entity name, e.g. 'DATASET'}

\item{con}{if NULL, return the name according to the package schema document. If provided, return the name used by the connection}
}
\value{
the namespace of the entity, e.g. 'gh_public'
}
\description{
return the namespace of an entity
}
