% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-schema.R
\name{full_arrayname}
\alias{full_arrayname}
\title{full name of array with namespace}
\usage{
full_arrayname(pkgEnv, entitynm, con = NULL)
}
\arguments{
\item{pkgEnv}{the package environment}

\item{entitynm}{an entity name, e.g. 'DATASET'}

\item{con}{if NULL, return the name according to the package schema document. If provided, return the name used by the connection}
}
\value{
the full name of the entity including namespace, e.g. 'gh_public.DATASET'
}
\description{
full name of array with namespace
}
