% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_entity}
\alias{get_entity}
\title{Get entity, optionally filtered by a dataframe of id(s)}
\usage{
get_entity(
  pkgEnv,
  con,
  entitynm,
  id = NULL,
  return_aop = FALSE,
  update_cache = F
)
}
\arguments{
\item{entitynm}{Entity name}

\item{id}{data.frame or arrayOp containing ID values to filter by}

\item{return_aop}{if FALSE (default), return a dataframe.  if TRUE, return an arrayop expression.}

\item{update_cache}{if entitynm is a cached entity (see \link{is_entity_cached}), force update the cache.  Applies only if \code{return_aop} is FALSE and \code{id} is NULL.  The cache is updated regardless if the previous condition holds and the cache for the entity is empty.}

\item{full_arrayname}{Full name of an array, or an AFL expression to filter on (e.g., something wrapped in secure_scan)}
}
\value{
arrayop query on entity or the corresponding dataframe
}
\description{
Get entity, optionally filtered by a dataframe of id(s)
}
