% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{glue_me}
\alias{glue_me}
\title{De-vectorized, NULL-immune \code{glue()}}
\usage{
glue_me(..., none = "(none)", .envir = parent.frame())
}
\arguments{
\item{...}{passed on to \code{glue::glue()}}

\item{none}{string value to use for NA and NULL}

\item{.envir}{execution environment}
}
\description{
De-vectorized, NULL-immune \code{glue()}
}
