% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{grouped_aggregate}
\alias{grouped_aggregate}
\title{Generate grouped_aggregate() expression}
\usage{
grouped_aggregate(aop, ..., fun = "count(*)", schema = NULL)
}
\arguments{
\item{aop}{arrayop object}

\item{...}{fields to group on}

\item{fun}{aggregate function}

\item{schema}{optional schema string to redimension into}
}
\value{
aop object
}
\description{
Generate afl expression for \code{grouped_aggregate(...)}
and \code{redimension(grouped_aggregate(...), ...)}
}
