% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_db}
\alias{init_db}
\title{Initialize or reinitialize arrays, roles, and namespaces associated with a REVEAL package}
\usage{
init_db(
  pkgEnv,
  con,
  init_arrays = TRUE,
  init_namespaces = FALSE,
  init_roles = FALSE,
  arrays_to_init = NULL,
  namespaces_to_init = NULL,
  roles_to_init = NULL,
  force = FALSE,
  silent = FALSE,
  allow_array_reinit = FALSE
)
}
\arguments{
\item{pkgEnv}{a REVEAL package environment object}

\item{con}{a connection object}

\item{init_arrays}{if TRUE, delete and initialize array.  Use with caution.}

\item{init_namespaces}{if TRUE, initialize namespaces.  Run only on first-time database initialization}

\item{init_roles}{if TRUE, initialize roles.  Run only on first-time database initialization}

\item{arrays_to_init}{(optional) if init_arrays is TRUE, list of arrays to delete and initialize.  if NULL (default), delete and initialize all arrays}

\item{namespaces_to_init}{(optional) if init_namespaces is TRUE, list of namespaces to initialize.  if NULL (default), initialize all namespaces.}

\item{roles_to_init}{(optional) if init_roles is TRUE, list of roles to initialize.  if NULL (default), initialize all roles.}

\item{force}{if TRUE, do not ask for confirmation prior to deletion / initialization}

\item{silent}{if TRUE, suppress messaging of individual namespace / role/ array deletion & initialization}

\item{allow_array_reinit}{if TRUE, allow deletion of arrays that are already initialized when init_arrays is TRUE.  if FALSE (default), initialize only arrays that do not yet exist.}
}
\description{
The function requires that you are connected to SciDB as an user with admin privileges
}
