% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_namespaces}
\alias{init_namespaces}
\title{Initialize namespaces associated with a REVEAL package}
\usage{
init_namespaces(pkgEnv, con, namespaces_to_init = NULL)
}
\description{
Initialize namespaces associated with a REVEAL package
}
