% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_roles}
\alias{init_roles}
\title{Initialize roles associated with a REVEAL package}
\usage{
init_roles(pkgEnv, con, roles_to_init = NULL)
}
\description{
Initialize roles associated with a REVEAL package
}
