% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{is_absolute}
\alias{is_absolute}
\title{Check whether or not \code{path} is absolute (i.e., starts with a \code{/})}
\usage{
is_absolute(path)
}
\arguments{
\item{path}{Path to check}
}
\value{
Logical indicating absolute path
}
\description{
Check whether or not \code{path} is absolute (i.e., starts with a \code{/})
}
