% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{list_files}
\alias{list_files}
\title{List "files" in file system or on S3}
\usage{
list_files(
  path = ".",
  pattern = NULL,
  full.names = FALSE,
  recursive = FALSE,
  ignore.case = FALSE,
  ...
)
}
\arguments{
\item{path}{File system path or S3 URL}

\item{pattern}{Filter pattern (regex)}

\item{full.names}{If set, return full path names rather than basenames}

\item{recursive}{List files recursively}

\item{ignore.case}{Ignore case for filter pattern}

\item{...}{Forwarded to \code{base::list.files()} or \code{aws.s3::get_bucket_df()} for files and S3 respectively}
}
\value{
Character vector of file names
}
\description{
Delegates to \code{base::list.files()} for file system paths and
mimics basic \code{base::list.files()} behavior for S3 paths.
}
