% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{list_users}
\alias{list_users}
\title{List all users}
\usage{
list_users(pkgEnv, con)
}
\description{
list the full list of users registered with scidb if the current user is an operator.  Otherwise, check a list of users that may be maintained by the package schema
}
