% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{map_or_echo}
\alias{map_or_echo}
\title{Map a string vector onto a set of new values from a list, but echo the input
for items not in the map}
\usage{
map_or_echo(map, mapee)
}
\arguments{
\item{map}{List mapping items names to item values}

\item{mapee}{Character vector to be mapped}
}
\value{
Character vector of mapped / echoed items
}
\description{
Map a string vector onto a set of new values from a list, but echo the input
for items not in the map
}
