% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{merge_find_matches_and_return_indices}
\alias{merge_find_matches_and_return_indices}
\title{helper function to merge multiple matches between vectors from \link{find_matches_and_return_indices}}
\usage{
merge_find_matches_and_return_indices(m1, m2)
}
\arguments{
\item{m1}{primary result from \link{find_matches_and_return_indices}}

\item{m2}{secondary match result, used for instances where m1 is unmatched}
}
\value{
list(match_res, source_matched_idx, source_unmatched_idx, target_matched_idx)
}
\description{
helper function to merge multiple matches between vectors from \link{find_matches_and_return_indices}
}
