% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{op_count}
\alias{op_count}
\title{Generate op_count() expression}
\usage{
op_count(aop)
}
\arguments{
\item{aop}{arrayop object}
}
\value{
aop object
}
\description{
Generate afl expression for \code{op_count(...)}
}
