% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p4logo}
\alias{p4logo}
\title{Get path to P4 logo image (transparent png, for use with light background)}
\usage{
p4logo()
}
\description{
Get path to P4 logo image (transparent png, for use with light background)
}
