% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{random_alphanum}
\alias{random_alphanum}
\title{Generate random alpha-numeric string}
\usage{
random_alphanum(prefix = "_", suffix = "", n = 3L)
}
\arguments{
\item{prefix}{prefix for random string}

\item{suffix}{suffix for random string}

\item{n}{length of string (excluding prefix and suffix)}
}
\description{
Generate random alpha-numeric string
}
