% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb.R
\name{remove_versions}
\alias{remove_versions}
\title{Remove versions of array after checking existence of versions}
\usage{
remove_versions(aop)
}
\arguments{
\item{aop}{arraop array to remove versions on}
}
\value{
reference to \code{aop}
}
\description{
Remove versions of array after checking existence of versions
}
