% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{s3read_using}
\alias{s3read_using}
\title{Run function on file downloaded from S3}
\usage{
s3read_using(
  FUN,
  object,
  ...,
  bucket,
  opts = NULL,
  filename = NULL,
  progress = TRUE
)
}
\arguments{
\item{FUN}{function accepting file path as first argument}

\item{object}{S3 object (i.e., object name or full s3://... url)}

\item{...}{passed as extra arguments to \code{FUN}}

\item{bucket}{S3 bucket, uses bucket from \code{object} if missing}

\item{opts}{see \code{aws.s3::s3read_using()}}

\item{filename}{File name to save S3 object to (defaults to using \code{tempfile()})}

\item{progress}{Toggle display of progress bar}
}
\description{
Modified version of \code{aws.s3::s3read_using()}. The difference with the original
implementation is that \code{filename} is taken "verbatim" rather than made relative
to tempdir.
}
