% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-schema.R
\name{scan_entity}
\alias{scan_entity}
\title{full name of array with namespace, wrapped in secure_scan if the current user needs secure_scan to read the array}
\usage{
scan_entity(pkgEnv, entitynm, con)
}
\arguments{
\item{pkgEnv}{the package environment}

\item{entitynm}{an entity name, e.g. 'DATASET'}

\item{con}{if NULL, return the name according to the package schema document. If provided, return the name used by the connection}
}
\value{
the full name of the entity including namespace and potentially wrapped in secure_scan, e.g. \code{gh_secure.DATASET} or \code{secure_scan(gh_secure.DATASET, strict:false)}
}
\description{
full name of array with namespace, wrapped in secure_scan if the current user needs secure_scan to read the array
}
