% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{set_permissions}
\alias{set_permissions}
\title{set secure_scan access permissions for a user}
\usage{
set_permissions(pkgEnv, con, user_id, secure_id, allowed, namespace = NULL)
}
\arguments{
\item{user_id}{the id of the user for whom to change permissions.  Obtained via \link{list_users}, or \code{-1} for all users (on Scidb 21.8 or above)}

\item{secure_id}{the ids along the secure dimension to change permissions for.}
}
\description{
to be run only by scidbadmin, or user
with Read / Write capability to \link{PERMISSIONS_ARRAY} and secured namespaces
}
\details{
can supply multiple ids at a time.
}
