% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{show_roles}
\alias{show_roles}
\title{Function to show roles associated with a schema}
\usage{
show_roles(pkgEnv, con = NULL)
}
\arguments{
\item{pkgEnv}{a REVEAL package environment object}

\item{con}{if not null, display whether or not role has been initialized in DB}
}
\description{
Function to show roles associated with a schema
}
