% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{show_roles_for_user}
\alias{show_roles_for_user}
\title{Function to show rules for user
Operator or higher privileges required to examine roles for a user other than oneself}
\usage{
show_roles_for_user(con, user_name = NULL)
}
\arguments{
\item{con}{connection object}

\item{user_name}{User name of user to show roles for.  If NULL, show roles for currently logged in user}
}
\description{
Function to show rules for user
Operator or higher privileges required to examine roles for a user other than oneself
}
