% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{show_user_namespace_permissions}
\alias{show_user_namespace_permissions}
\title{Function to show what permissions a user has to a namespace according to the schema}
\usage{
show_user_namespace_permissions(pkgEnv, con, user_name = NULL, namespace)
}
\arguments{
\item{con}{a connection object}

\item{user_name}{the user for whom to show permissions.  If \code{NULL} (default), use the current logged in user}

\item{namespace}{the namespace for which to show permissions}
}
\value{
a list with names \code{c("c", "r", "u", "l", "d")} with value \code{TRUE} if the user has the corresponding permission in the specified namespace
}
\description{
Function to show what permissions a user has to a namespace according to the schema
}
