% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{show_users_in_role}
\alias{show_users_in_role}
\title{Show users with a specified role
Run only with operator or higher privileges}
\usage{
show_users_in_role(con, role)
}
\arguments{
\item{con}{connection object (optional when using \code{rg_connect()})}

\item{role}{Name of role to generate list of users for}
}
\description{
Show users with a specified role
Run only with operator or higher privileges
}
