% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{uri_download}
\alias{uri_download}
\title{Download file from uri}
\usage{
uri_download(
  uri,
  filename = tempfile(fileext = file_ext(uri)),
  ...,
  progress = TRUE
)
}
\arguments{
\item{uri}{URI (i.e., file path or S3 or http(s) URL) specifying source location}

\item{filename}{File name (path) to store file at (defaults to \code{tempfile()})}

\item{...}{Forwarded to internal download functions}

\item{progress}{If set, show download progress while downloading}
}
\value{
Character vector with file path where file was downloaded
}
\description{
Downloads file from S3 or http(s) to a file and returned the file name. If a file path is
specified, the path will be returned if the file exists.
}
