% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{check_user_admin_status}
\alias{check_user_admin_status}
\title{Function to check if current user has admin permissions}
\usage{
check_user_admin_status(con, user_name = get_logged_in_user(con))
}
\arguments{
\item{con}{a connection object}
}
\description{
Function to check if current user has admin permissions
}
