% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_arrays}
\alias{init_arrays}
\title{Delete and initialize arrays associated with a REVEAL package}
\usage{
init_arrays(
  pkgEnv,
  con,
  arrays_to_init = NULL,
  allow_array_reinit = FALSE,
  force = FALSE,
  silent = FALSE
)
}
\description{
Delete and initialize arrays associated with a REVEAL package
}
